refactor deployment
