+++ backend/app/tasks/deploy_frame.py
@@ -1,14 +1,9 @@
 from datetime import datetime, timezone
 import json
-import math
 import os
-from pathlib import Path
-import random
 import hashlib
 import re
-import shlex
 import shutil
-import string
 import tempfile
 from typing import Any, Optional
 
@@ -23,74 +18,18 @@ from app.drivers.waveshare import write_waveshare_driver_nim, get_variant_folder
 from app.models import get_apps_from_scenes
 from app.models.assets import sync_assets
 from app.models.log import new_log as log
-from app.models.frame import Frame, update_frame, get_frame_json
-from app.utils.remote_exec import run_command, upload_file
+from app.models.frame import Frame, update_frame
+from app.utils.remote_exec import upload_file
 from app.utils.local_exec import exec_local_command
 from app.models.apps import get_one_app_sources
 from app.models.settings import get_settings_dict
 from app.utils.nix_utils import nix_attr, nix_cmd
+from app.tasks._frame_deployer import FrameDeployer
 
 from .utils import find_nim_v2, find_nimbase_file
 
 async def deploy_frame(id: int, redis: Redis):
     await redis.enqueue_job("deploy_frame", id=id)
-
-class FrameDeployer:
-    ## This entire file is a big refactor in progress.
-    def __init__(self, db: Session, redis: Redis, frame: Frame, nim_path: str, temp_dir: str):
-        self.db = db
-        self.redis = redis
-        self.frame = frame
-        self.nim_path = nim_path
-        self.temp_dir = temp_dir
-        self.build_id = ''.join(random.choice(string.ascii_lowercase) for _ in range(12))
-        self.deploy_start: datetime = datetime.now()
-
-    async def exec_command(
-        self,
-        command: str,
-        output: Optional[list[str]] = None,
-        log_output: bool = True,
-        log_command: str | bool = True,
-        raise_on_error: bool = True,
-        timeout: int = 600 # 10 minutes default timeout
-    ) -> int:
-        status, stdout, stderr = await run_command(
-            self.db, self.redis, self.frame, command, log_output=log_output, log_command=log_command, timeout=timeout
-        )
-        if output is not None:
-            lines = (stdout + "\n" + stderr).splitlines()
-            output.extend(lines)
-        if status != 0 and raise_on_error:
-            raise Exception(
-                f"Command '{command}' failed with code {status}\nstderr: {stderr}\nstdout: {stdout}"
-            )
-        return status
-
-    async def log(self, type: str, line: str, timestamp: Optional[datetime] = None):
-        await log(self.db, self.redis, int(self.frame.id), type=type, line=line, timestamp=timestamp)
-
-    async def _store_paths_missing(self, paths: list[str]) -> list[str]:
-        """
-        Return *only* those paths that are **missing** from /nix/store on the
-        frame.  All paths are checked in a single SSH exec.
-        """
-        if not paths:
-            return []
-
-        # Quote every path once; the remote loop echoes the missing ones.
-        joined = " ".join(shlex.quote(p) for p in paths)
-        script = "for p; do [ -e \"$p\" ] || echo \"$p\"; done"
-        out: list[str] = []
-        await self.exec_command(
-            f"bash -s -- {joined} <<'EOF'\n{script}\nEOF",
-            output=out,
-            raise_on_error=False,
-            log_command=f"bash -s -- **SKIPPED** <<'EOF'\n{script}\nEOF",
-            log_output=False,
-        )
-        # Every line produced by the loop is a missing store path.
-        return [p.strip() for p in out if p.strip()]
 
 async def deploy_frame_task(ctx: dict[str, Any], id: int):
     db: Session = ctx['db']
@@ -145,94 +84,63 @@ async def deploy_frame_task(ctx: dict[str, Any], id: int):
             await self.log("stdout", f"Deploying frame {frame.name} with build id {self.build_id}")
 
             await self.log("stdout", "- Getting target architecture")
-            uname_output: list[str] = []
-            await self.exec_command("uname -m", uname_output)
-            arch = "".join(uname_output).strip()
-            if arch in ("aarch64", "arm64"):
-                cpu = "arm64"
-            elif arch in ("armv6l", "armv7l"):
-                cpu = "arm"
-            elif arch == "i386":
-                cpu = "i386"
-            else:
-                cpu = "amd64"
 
-            distro_out: list[str] = []
-            await self.exec_command(
-                "bash -c '"
-                "if [ -e /etc/nixos/version ]; then echo nixos ; "
-                "elif [ -f /etc/rpi-issue ] || grep -q \"^ID=raspbian\" /etc/os-release ; then echo raspios ; "
-                "else . /etc/os-release ; echo ${ID:-unknown} ; "
-                "fi'",
-                distro_out
-            )
-            distro = distro_out[0].strip().lower()
-
-            mem_output: list[str] = []
-            await self.exec_command(
-                "grep MemTotal /proc/meminfo | awk '{print $2}'",
-                mem_output,
-            )
-            kib = int(mem_output[0].strip()) if mem_output else 0   # kB from the kernel
-            total_memory = kib // 1024                             # MiB
+            arch = await self.get_cpu_architecture()
+            distro = await self.get_distro()
+            total_memory = await self.get_total_memory_mb()
             low_memory = total_memory < 512
 
             await self.log("stdout", f"- Detected distro: {distro}, architecture: {arch}, total memory: {total_memory} MiB")
 
-            with tempfile.TemporaryDirectory() as temp_dir_local:
-                await self.log("stdout", f"- Preparing sources with local modifications under {temp_dir_local}")
-                source_dir_local = create_local_source_folder(temp_dir_local)
-                await make_local_modifications(self, source_dir_local)
+            await self.log("stdout", f"- Preparing sources with local modifications under {temp_dir}")
+            source_dir_local = create_local_source_folder(temp_dir)
+            await make_local_modifications(self, source_dir_local)
 
-                # Fast-path for NixOS targets
-                if distro == "nixos":
-                    await self.log("stdout", "- NixOS detected – using flake-based deploy")
+            # Fast-path for NixOS targets
+            if distro == "nixos":
+                await self.log("stdout", "- NixOS detected – using flake-based deploy")
 
-                    target_host = await get_hostname(self)
-                    sys_build_cmd, masked_build_cmd, cleanup = nix_cmd(
-                        "nix --extra-experimental-features 'nix-command flakes' "
-                        f"build \"$(realpath {source_dir_local})\"#nixosConfigurations.{nix_attr(target_host)}.config.system.build.toplevel "
-                        "--system aarch64-linux --print-out-paths "
-                        "--log-format raw -L",
-                        settings
+                target_host = await get_hostname(self)
+                sys_build_cmd, masked_build_cmd, cleanup = nix_cmd(
+                    "nix --extra-experimental-features 'nix-command flakes' "
+                    f"build \"$(realpath {source_dir_local})\"#nixosConfigurations.{nix_attr(target_host)}.config.system.build.toplevel "
+                    "--system aarch64-linux --print-out-paths "
+                    "--log-format raw -L",
+                    settings
+                )
+                try:
+                    await self.log("stdout", f"$ {masked_build_cmd}")
+                    status, sys_out, sys_err = await exec_local_command(self.db, self.redis, self.frame, sys_build_cmd, log_command=False)
+                finally:
+                    cleanup()
+                if status != 0 or not sys_out:
+                    raise Exception(f"Local NixOS system build failed:\n{sys_err}")
+                result_path = sys_out.strip().splitlines()[-1]
+
+                updated_count = await self.nix_upload_path_and_deps(result_path)
+
+                await self._upload_frame_json("/var/lib/frameos/frame.json")
+
+                if updated_count == 0:
+                    await self.restart_service("frameos")
+                else:
+                    await self.exec_command(f"sudo nix-env --profile /nix/var/nix/profiles/system --set {result_path}")
+                    await self.exec_command(
+                        "sudo systemd-run --unit=nixos-switch --no-ask-password "
+                        "/nix/var/nix/profiles/system/bin/switch-to-configuration switch"
                     )
-                    try:
-                        await self.log("stdout", f"$ {masked_build_cmd}")
-                        status, sys_out, sys_err = await exec_local_command(self.db, self.redis, self.frame, sys_build_cmd, log_command=False)
-                    finally:
-                        cleanup()
-                    if status != 0 or not sys_out:
-                        raise Exception(f"Local NixOS system build failed:\n{sys_err}")
-                    result_path = sys_out.strip().splitlines()[-1]
-
-                    updated_count = await nix_upload_path_and_deps(self, result_path)
-
-                    frame_json_data = (json.dumps(get_frame_json(db, frame), indent=4) + "\n").encode("utf-8")
-                    await upload_file(self.db, self.redis, self.frame, "/var/lib/frameos/frame.json", frame_json_data)
-
-                    if updated_count == 0:
-                        await self.exec_command("sudo systemctl enable frameos.service")
-                        await self.exec_command("sudo systemctl restart frameos.service")
-                        await self.exec_command("sudo systemctl status frameos.service")
-                    else:
-                        await self.exec_command(f"sudo nix-env --profile /nix/var/nix/profiles/system --set {result_path}")
-                        await self.exec_command(
-                            "sudo systemd-run --unit=nixos-switch --no-ask-password "
-                            "/nix/var/nix/profiles/system/bin/switch-to-configuration switch"
-                        )
-                    #  Save deploy metadata & finish early – nothing else to do
-                    frame.status = 'starting'
-                    frame.last_successful_deploy     = frame_dict
-                    frame.last_successful_deploy_at  = datetime.now(timezone.utc)
-                    await update_frame(db, redis, frame)
-                    await self.log("stdinfo", f"Deploy finished in {datetime.now() - self.deploy_start} 🎉")
-                    return   # ← all done, skip the legacy RPiOS flow
+                #  Save deploy metadata & finish early – nothing else to do
+                frame.status = 'starting'
+                frame.last_successful_deploy     = frame_dict
+                frame.last_successful_deploy_at  = datetime.now(timezone.utc)
+                await update_frame(db, redis, frame)
+                await self.log("stdinfo", f"Deploy finished in {datetime.now() - self.deploy_start} 🎉")
+                return   # ← all done, skip the legacy RPiOS flow
 
             ## /END NIXOS
 
 
             ## Deploy onto Raspberry Pi OS or Debian/Ubuntu:
-
 
             if distro == "raspios":
                 await self.log("stdout", "- Raspberry Pi OS detected")
@@ -251,7 +159,7 @@ async def deploy_frame_task(ctx: dict[str, Any], id: int):
             await self.log("stdout", "- Applying local modifications")
             await make_local_modifications(self, source_dir)
             await self.log("stdout", "- Creating build archive")
-            archive_path = await create_local_build_archive(self, build_dir, source_dir, cpu)
+            archive_path = await create_local_build_archive(self, build_dir, source_dir, arch)
 
             if low_memory:
                 await self.log("stdout", "- Low memory device, stopping FrameOS for compilation")
@@ -356,15 +264,7 @@ async def deploy_frame_task(ctx: dict[str, Any], id: int):
             )
 
             # 4. Upload frame.json
-            frame_json_data = (
-                json.dumps(get_frame_json(db, frame), indent=4) + "\n"
-            ).encode("utf-8")
-            await upload_file(
-                self.db, self.redis, self.frame,
-                f"/srv/frameos/releases/release_{build_id}/frame.json",
-                frame_json_data,
-            )
-            await self.log("stdout", f"> add /srv/frameos/releases/release_{build_id}/frame.json")
+            await self._upload_frame_json(f"/srv/frameos/releases/release_{build_id}/frame.json")
 
             # Driver-specific vendor steps
             if inkyPython := drivers.get("inkyPython"):
@@ -485,9 +385,6 @@ async def deploy_frame_task(ctx: dict[str, Any], id: int):
                     await self.exec_command(command=f'echo "{line}" | sudo tee -a ' + boot_config, log_output=False)
                     must_reboot = True
 
-        await self.exec_command("sudo systemctl daemon-reload")
-        await self.exec_command("sudo systemctl enable frameos.service")
-
         if frame.last_successful_deploy_at is None:
             # Reboot after the first deploy to make sure any modifications to config.txt are persisted to disk
             # Otherwise if you pull out the power, you'll end up with a blank config.txt on the next boot.
@@ -505,8 +402,8 @@ async def deploy_frame_task(ctx: dict[str, Any], id: int):
             await self.log("stdinfo", "Deployed! Rebooting device after boot config changes")
             await self.exec_command("sudo reboot")
         else:
-            await self.exec_command("sudo systemctl restart frameos.service")
-            await self.exec_command("sudo systemctl status frameos.service")
+            await self.exec_command("sudo systemctl daemon-reload")
+            await self.restart_service("frameos")
             await update_frame(db, redis, frame)
 
     except Exception as e:
@@ -584,7 +481,7 @@ async def create_local_build_archive(
     self: FrameDeployer,
     build_dir: str,
     source_dir: str,
-    cpu: str
+    arch: str
 ):
     db = self.db
     redis = self.redis
@@ -619,6 +516,7 @@ async def create_local_build_archive(
     await self.log("stdout",
               "- No cross compilation. Generating source code for compilation on frame.")
 
+    cpu = await self.arch_to_nim_cpu(arch)
     debug_options = "--lineTrace:on" if frame.debug else ""
     cmd = (
         f"cd {source_dir} && nimble assets -y && nimble setup && "
@@ -717,118 +615,6 @@ async def create_local_build_archive(
     zip_base = os.path.join(temp_dir, f"build_{build_id}")
     shutil.make_archive(zip_base, 'gztar', temp_dir, f"build_{build_id}")
     return archive_path
-
-BYTES_PER_MB   = 1_048_576
-DEFAULT_CHUNK  = 25 * BYTES_PER_MB
-
-async def nix_upload_path_and_deps(
-    self: "FrameDeployer",
-    path: str,
-    max_chunk_size: int = DEFAULT_CHUNK,
-) -> int: # return number of uploaded items
-    """
-    Export the full runtime closure of *path* and import it on the target
-    machine, but bundle the nar streams so that at most `max_chunk_size`
-    bytes are transferred per upload (≈10 MiB by default).
-
-    The implementation relies on
-        • `nix path-info --json` to get the *narSize* of every path once,
-        • `nix-store --export …` which can export many store paths in one
-          go, producing a concatenated nar stream,
-        • `nix-store --import` on the device to unpack an entire chunk.
-    """
-    await self.log("stdout", f"- Collecting runtime closure for {path}")
-
-    # 1. Get complete closure
-    status, paths_out, err = await exec_local_command(
-        self.db, self.redis, self.frame, f"nix-store -qR {path}", log_output=False
-    )
-    if status:
-        raise RuntimeError(f"Failed to collect closure: {err}")
-
-    runtime_paths = (paths_out or "").strip().splitlines()
-    await self.log("stdout", f"  → {len(runtime_paths)} store paths")
-
-    # 2. Filter out paths that are already present on the device
-    missing = await self._store_paths_missing(runtime_paths)
-    if not missing:
-        await self.log("stdout", "  → No missing store paths, skipping upload")
-        return 0
-    await self.log(
-        "stdout",
-        f"  → {len(missing)} paths need upload; bundling in ≤{max_chunk_size // BYTES_PER_MB} MiB chunks"
-    )
-
-    # 3. Query nar sizes once for all paths
-    cmd = ["nix", "path-info", "--json", *missing]
-    status, size_json, err = await exec_local_command(
-        self.db, self.redis, self.frame, " ".join(cmd), log_output=False
-    )
-    if status:
-        raise RuntimeError(f"nix path-info failed: {err}")
-
-    size_info: dict[str, int] = {}
-    info = json.loads(size_json or "{}")
-
-    if isinstance(info, dict):
-        # `info` maps path → metadata
-        for p, meta in info.items():
-            # prefer `narSize` (current), fall back to legacy `nar`
-            size_info[p] = int(meta.get("narSize") or meta.get("nar") or 0)
-    else:  # older nix (<2.4) returned a list
-        for meta in info:                          # type: ignore[arg-type]
-            size_info[meta["path"]] = int(meta.get("narSize") or meta.get("nar") or 0)
-
-    # 4. Greedily pack paths up to ~max_chunk_size each
-    chunks: list[list[str]] = []
-    current: list[str] = []
-    current_size = 0
-    for p in missing:
-        nar_size = size_info.get(p, 0)
-        # start new chunk if adding would overflow (but always put at least one)
-        if current and current_size + nar_size > max_chunk_size:
-            chunks.append(current)
-            current, current_size = [], 0
-        current.append(p)
-        current_size += nar_size
-    if current:
-        chunks.append(current)
-
-    await self.log("stdout", f"  → Uploading in {len(chunks)} chunk(s)")
-
-    remote_tmp = f"/tmp/frameos_import_{self.build_id}"
-    await self.exec_command(f"mkdir -p {remote_tmp}")
-
-    try:
-        for i, chunk in enumerate(chunks, 1):
-            with tempfile.TemporaryDirectory() as tmpdir:
-                nar_local = Path(tmpdir) / f"chunk_{i}.nar"
-                export_cmd = f"nix-store --export {' '.join(shlex.quote(p) for p in chunk)} > {nar_local}"
-                status, _, err = await exec_local_command(
-                    self.db, self.redis, self.frame, export_cmd
-                )
-                if status:
-                    raise RuntimeError(f"Export failed for chunk {i}: {err}")
-
-                # 5. Ship and import the chunk
-                remote_nar = f"{remote_tmp}/chunk_{i}.nar"
-                with open(nar_local, "rb") as fh:
-                    await upload_file(
-                        self.db, self.redis, self.frame, remote_nar, fh.read()
-                    )
-
-                await self.exec_command(
-                    f"sudo nix-store --import < {remote_nar} && rm {remote_nar}"
-                )
-                await self.log(
-                    "stdout",
-                    f"🍀 imported chunk {i}/{len(chunks)} 🍀 "
-                    f"({len(chunk)} paths, {math.ceil(nar_local.stat().st_size/BYTES_PER_MB)} MiB)"
-                )
-    finally:
-        await self.exec_command(f"rmdir {remote_tmp}")
-
-    return len(missing)
 
 async def file_in_sync(
     self: "FrameDeployer",
