+++ backend/app/tasks/fast_deploy_frame.py
@@ -1,13 +1,12 @@
 from datetime import datetime, timezone
-import json
 from typing import Any
 
 from sqlalchemy.orm import Session
 from arq import ArqRedis as Redis
 
 from app.models.log import new_log as log
-from app.models.frame import Frame, update_frame, get_frame_json
-from app.utils.remote_exec import run_commands, upload_file
+from app.models.frame import Frame, update_frame
+from app.tasks._frame_deployer import FrameDeployer
 
 async def fast_deploy_frame(id: int, redis: Redis):
     await redis.enqueue_job("fast_deploy_frame", id=id)
@@ -26,12 +25,9 @@ async def fast_deploy_frame_task(ctx: dict[str, Any], id: int):
         frame.status = "restarting"
         await update_frame(db, redis, frame)
 
-        await run_commands(
-            db,
-            redis,
-            frame,
-            ["sudo systemctl stop frameos.service || true"],
-        )
+        self = FrameDeployer(db=db, redis=redis, frame=frame, nim_path="", temp_dir="")
+
+        await self.stop_service('frameos')
 
         frame_dict = frame.to_dict() # persisted as frame.last_successful_deploy if successful
         if "last_successful_deploy" in frame_dict:
@@ -39,27 +35,13 @@ async def fast_deploy_frame_task(ctx: dict[str, Any], id: int):
         if "last_successful_deploy_at" in frame_dict:
             del frame_dict["last_successful_deploy_at"]
 
-        frame_json_data = (
-            json.dumps(get_frame_json(db, frame), indent=4).encode() + b"\n"
-        )
-        await upload_file(
-            db,
-            redis,
-            frame,
-            "/srv/frameos/current/frame.json",
-            frame_json_data,
-        )
+        arch = await self.get_cpu_architecture()
+        if arch == 'nixos':
+            await self._upload_frame_json("/var/lib/frameos/frame.json")
+        else:
+            await self._upload_frame_json("/srv/frameos/current/frame.json")
 
-        await run_commands(
-            db,
-            redis,
-            frame,
-            [
-                "sudo systemctl enable frameos.service",
-                "sudo systemctl start frameos.service",
-                "sudo systemctl status frameos.service",
-            ],
-        )
+        await self.restart_service('frameos')
 
         frame.status = 'starting'
         frame.last_successful_deploy = frame_dict
